requirements.txt
gradio>=3.40.1
pypandoc
pdf2docx
python-docx
flask
python-multipart